Create new packages for datetimeAlgorithm and hashAlgorithm
